Fix: factorial for large integers

+++ main.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
-int factorial(int num)
+long factorial(int num)
 {
-    int f = 1;
+    long f = 1;
     for (int i = 2; i <= num; i++)
         f = f * i;
     return f;
@@ -10,6 +10,6 @@ int factorial(int num)
 
 int main()
 {
-    printf("Factorial = %d", factorial(5));
+    printf("Factorial = %ld", factorial(5));
     return 0;
 }
